backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
Pillow==11.3.0
aiofiles==23.1.0
numpy==1.26.4
opencv-python==4.10.0.84
